✨ Register Minigame Modules
Now the registration of new minigame instances is not done by registering an Assembly, this method was breaking the use of minigames that were in other Assemblies, so a new function was added in the MinigameInstanceClient class, it is called RegisterMinigameModule<T>() and allows you to register a minigame module.
This will allow the registration of several minigames and not just the ones that are in a single Assembly.